Code revision
Updated to Python 3 and made the loops more pythonic.

Added a second dictionary to hold the frequencies:
The previous code did not actually work when the model was trained multiple times. In fact, the old probabilities (values between 0.0 and 1.0) were overwritten when updating them with the frequencies (values from 1 to ...).
Now there are two dictionaries: one containing the frequencies and one the probabilities.
At every training the former is used to increase the counts, the latter is discarded and recomputed.

Improvements:
I'm no expert but a dictionary of dictionaries might not be the fastest data structure in the world. I've started coding a similar project using ordered lists with binary search and numpy's matrices for faster computations, have a look at https://github.com/baldassarreFe/python-pankakes, comments are welcome ;)